feat: Add LayoutTopNav component and update tsconfig.json
