+++ frontend/occupi-web/src/components/index.ts
@@ -1,0 +1,43 @@
+import DrawerComponent from "./drawerComponent/DrawerComponent";
+import GradientButton from "./gradientButtonComponent/gradientButton";
+import { Checkbox } from "./ui/checkbox";
+import InputBox from "./InputBox/InputBox";
+import OtpComponent from "./OtpComponent/OtpComponent";
+import SideNavBarButton from "./SideNavBarButton/SideNavBarButton";
+import TopNav from './topNav/TopNav';
+import GraphContainer from "./graphContainer/GraphContainer";
+import TabComponent from "./tabComponent/TabComponent";
+import ColourCircle from "./colourCircle/ColourCircle";
+import Appearance from "./appearance/Appearance";
+import MenuItem from './drawerComponent/menuItem/MenuItem';
+import Bookings from "./overView/OverView";
+import OverView from "./bookings/Bookings";
+import OverviewComponent from "./OverviewComponent/OverviewComponent";
+import ProfileComponent from "./ProfileComponent/ProfileComponent";
+import SearchBarComponent from "./searchBarComponent/SearchBar";
+import SideNav from "./sideNavComponent/SideNav";
+
+
+
+export {
+    DrawerComponent,
+    GradientButton,
+    Checkbox,
+    InputBox,
+    OtpComponent,
+    SideNavBarButton,
+    TopNav,
+    GraphContainer,
+    TabComponent,
+    ColourCircle,
+    Appearance,
+    MenuItem,
+    Bookings,
+    OverView,
+    OverviewComponent,
+    ProfileComponent,
+    SearchBarComponent,
+    SideNav,
+
+
+}
