+++ frontend/occupi-web/src/layouts/index.ts
@@ -1,5 +1,6 @@
 import Layout from './Layout'
+import LayoutTopNav from './LayoutTopNav'
 
 export {
-    Layout
+    Layout,LayoutTopNav
 }
