+++ frontend/occupi-web/src/pages/index.ts
@@ -4,7 +4,7 @@ import LandingPage from './landing-page/LandingPage'
 import Settings from './settings-page/Settings'
 import Dashboard from "./Dashboard/Dashboard";
 import Appearance from './appearance-page/Appearance';
-import OverView from '../components/overView/OverView';
+import OverView from '../components/bookings/Bookings';
 export {
     LoginForm, OtpPage, LandingPage, Settings,
     Dashboard,Appearance ,OverView
